fix(comvis): unsignedBigInteger -> unsignedInteger em business_id (FK type mismatch) (#477)
4 migrations comvis_* declaravam business_id como unsignedBigInteger mas
business.id (UltimatePOS legacy) e int(10) unsigned. Type mismatch
quebra FK constraint em prod MySQL com FK strict mode:

  SQLSTATE[HY000]: General error: 1005 Can't create table
  `u906587222_oimpresso`.`comvis_materiais` (errno: 150
  "Foreign key constraint is incorrectly formed")

Local dev pode ter passado por SQLite (sem FK strict) ou MySQL config
diferente. Hostinger prod bloqueou todo migrate em sequencia, deixando
2026_05_10_120000_alter_nfe_emissoes_status_enum_add_enviando_erro_envio
PENDING — bug critico pq pipeline NFC-e PR #434 precisa do ENUM novo.

Fix:
- 2026_05_10_000040_create_comvis_materiais: 1 troca
- 2026_05_10_000041_create_comvis_orcamentos: 2 trocas (orcamentos +
  orcamento_itens redundante pro global scope)
- 2026_05_10_000042_create_comvis_os: 1 troca
- 2026_05_10_000043_create_comvis_apontamentos: 1 troca

Total: 5 trocas em 4 arquivos. Outras FKs (vendedor_id, operador_id,
contato_id) ja eram unsignedInteger correto. FKs entre tabelas comvis_*
mantem unsignedBigInteger (todas usam bigIncrements).

Comentario adicionado em cada coluna citando UltimatePOS legacy schema
pra prevenir reincidencia.

Refs: deploy 2026-05-10 pos-merge auditoria · FK constraint MySQL strict
desbloqueia migration ENUM critica pra #434

+++ Modules/ComunicacaoVisual/Database/Migrations/2026_05_10_000043_create_comvis_apontamentos_table.php
@@ -24,7 +24,8 @@ return new class extends Migration {
     {
         Schema::create('comvis_apontamentos', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedBigInteger('business_id');
+            // int(10) unsigned pra bater com business.id (UltimatePOS legacy schema — FK constraint)
+            $table->unsignedInteger('business_id');
             $table->unsignedBigInteger('os_id');                    // FK comvis_os
             $table->unsignedBigInteger('orcamento_item_id')->nullable(); // FK comvis_orcamento_itens — null = OS sem item específico
             $table->unsignedInteger('operador_id');                 // FK users.id — quem apontou
